+++ playwright/tests/keywordSearch.spec.ts
@@ -1,17 +1,9 @@
-import { Given, Then, When } from "@cucumber/cucumber";
+import { Then, When } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
 
-Given(
-  "I am on the DRB home page for searching",
-  async function (this: CustomWorld) {
-    return await this.page.goto(`${this.parameters.appUrl}`);
-  }
-);
-
 When("I input a search term", async function (this: CustomWorld) {
   await this.page.locator('[aria-label="Item Search"]').fill("IBM 1401");
-  await this.page.locator("#searchbar-button-search-bar").click();
 });
 
 Then(
@@ -22,6 +14,5 @@ Then(
       .waitFor();
     const textIWant = this.page.locator("a", { hasText: "IBM 1401" });
     expect(await textIWant.count()).toBeGreaterThan(3);
-    console.log(textIWant);
   }
 );