+++ playwright/tests/pub_yearFilter.spec.ts
@@ -1,15 +1,9 @@
-import { Given, Then, When } from "@cucumber/cucumber";
+import { Then, When } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
 
-Given("Given I am on the home page", async function (this: CustomWorld) {
-  return await this.page.goto(`${this.parameters.appUrl}`);
-});
-
 When("When I search for New York", async function (this: CustomWorld) {
   await this.page.locator('[aria-label="Item Search"]').fill("New York");
-  await this.page.locator("#searchbar-button-search-bar").click();
-  // await this.page.getByRole("button", { name: /submit/i }).click();
 });
 
 Then(
@@ -17,7 +11,6 @@ Then(
   async function (this: CustomWorld) {
     await this.page.locator('[name="Date From"]').fill("1900");
     await this.page.locator('[name="Date To"]').fill("1900");
-    await this.page.locator("#year-filter-button").click();
   }
 );