+++ playwright/tests/titleSearch.spec.ts
@@ -1,10 +1,7 @@
-import { Given, Then, When } from "@cucumber/cucumber";
+import { Then, When } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
 
-Given("I want to do a title search", async function (this: CustomWorld) {
-  return await this.page.goto("https://digital-research-books-beta.nypl.org/");
-});
 When("I change the dropdown to title", async function (this: CustomWorld) {
   const dropdown = await this.page.$("[aria-label='Select a search category']");
   await dropdown.selectOption({ value: "title" });
@@ -12,7 +9,6 @@ When("I change the dropdown to title", async function (this: CustomWorld) {
 
 When("I input a title search term", async function (this: CustomWorld) {
   await this.page.locator("[aria-label='Item Search']").fill("IBM 1401");
-  await this.page.locator("#searchbar-button-search-bar").click();
 });
 
 Then(