+++ src/components/About/About.test.tsx
@@ -1,12 +1,8 @@
 import React from "react";
 import AboutPage from "./About";
-import renderer from "react-test-renderer";
-import { createSerializer } from "@emotion/jest";
-jest.mock("next/router", () => require("next-router-mock"));
-
-expect.addSnapshotSerializer(createSerializer());
+import { render } from "@testing-library/react";
 
 it("renders about unchanged", async () => {
-  const tree = await renderer.create(<AboutPage />).toJSON();
-  expect(tree).toMatchSnapshot();
+  const tree = render(<AboutPage />);
+  expect(tree.container.firstChild).toMatchSnapshot();
 });