+++ src/components/AdvancedSearch/AdvancedSearch.tsx
@@ -24,7 +24,6 @@ import {
   Button,
   ButtonGroup,
   Checkbox,
-  Footer,
   Form,
   FormField,
   FormRow,
@@ -272,8 +271,8 @@ const AdvancedSearch: React.FC<{
             </FormField>
             <FormField>
               <FilterYears
-                startFilter={startFilter && startFilter[0]}
-                endFilter={endFilter && endFilter[0]}
+                startFilter={startFilter}
+                endFilter={endFilter}
                 onDateChange={(
                   e: React.ChangeEvent<HTMLInputElement>,
                   isStart: boolean
@@ -315,9 +314,7 @@ const AdvancedSearch: React.FC<{
           </Form>
         </TemplateContentPrimary>
       </TemplateContent>
-      <TemplateFooter>
-        <Footer />
-      </TemplateFooter>
+      <TemplateFooter />
     </Template>
   );
 };